+++ src/edge.ts
@@ -1,49 +1,56 @@
 import { NodeId } from './node';
 import _ from 'lodash';
-import { Serializable } from './index';
 
 export type EdgeId = string;
 
-export type RawEdge<ED extends Serializable> = {
+export type RawEdge<E> = {
   sourceId: NodeId, 
   targetId: NodeId, 
-  attr: ED
+  attr: E
 }
 
 /**
  * A single directed edge consisting of a source id, target id,
  * and the data associated with the edge.
  *
- * @tparam ED type of the edge attribute
+ * @tparam E type of the edge attribute
  *
  * @param srcId The vertex id of the source vertex
  * @param dstId The vertex id of the target vertex
  * @param attr The attribute associated with the edge
  */
-export class Edge<ED extends Serializable> {
-  constructor(readonly sourceId: NodeId, readonly targetId: NodeId, readonly attr: ED) {
+export class GraphEdge<E> {
+  constructor(readonly sourceId: NodeId, readonly targetId: NodeId, readonly attr: E) {
     this.sourceId = sourceId;
     this.targetId = targetId;
     this.attr = attr;
   }
 
   toString() {
+    let attrStr: string = '';
+    if (!!this.attr['toString'] && typeof this.attr['toString'] === 'function'){
+      //@ts-ignore
+      attrStr = this.attr.toString();
+    }
+    else{
+      attrStr = JSON.stringify(this.attr);
+    }
     return JSON.stringify(
       {
         sourceId: this.sourceId,
         targetId: this.targetId, 
-        attr: this.attr.toString()
+        attr: attrStr
       }
     );
   }
 
   static fromObject(obj:{ sourceId: string, targetId: string, attr: any }) {
-    return new Edge(obj.sourceId, obj.targetId, obj.attr);
+    return new GraphEdge(obj.sourceId, obj.targetId, obj.attr);
   }
 
   static fromString(json: string) {
     const obj = JSON.parse(json);
-    return new Edge(obj.sourceId, obj.targetId, obj.attr);
+    return new GraphEdge(obj.sourceId, obj.targetId, obj.attr);
   }
 
   static edgeId(sourceId: NodeId, targetId: NodeId): EdgeId {