+++ src/json.ts
@@ -1,6 +1,6 @@
 import { Graph } from './index'
 import _ from 'lodash';
-import { Serializable } from './index';
+
 
 export function toJson(graph: Graph<any, any>){
     let nodes = graph.nodes;