+++ src/serializable.ts
@@ -1,7 +1,0 @@
-
-export interface Serializable {
-    toString(): string
-    fromString?(json:string): any
-}
-
-